app/pyproject.toml
[project]
name = "telegram-nutrition-assistant"
version = "0.1.0"
description = "Telegram Nutrition Assistant"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.12.0",
    "cachetools==6.2.2",
    "certifi==2025.11.12",
    "charset-normalizer==3.4.4",
    "distro==1.9.0",
    "google-auth==2.43.0",
    "google-auth-oauthlib==1.2.2",
    "google-genai==1.55.0",
    "gspread==6.2.1",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.11",
    "oauthlib==3.3.1",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "python-dotenv==1.2.1",
    "python-telegram-bot==22.5",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "rsa==4.9.1",
    "sniffio==1.3.1",
    "tenacity==9.1.2",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.6.1",
    "websockets==15.0.1",
]

